Added eslint. Restructured project
