+++ .vs/isaf.code-workspace
@@ -3,6 +3,8 @@
 		"path": ".."
 	}],
 	"settings": {
+		"editor.tabSize": 2,
+		"editor.insertSpaces": false,
 		"files.exclude": {
 			"node_modules": true,
 			"package-lock.json": true
